requirements.txt
nicegui>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-dotenv>=1.0.0
numpy>=2.0.0
scikit-learn>=1.5.0
pydantic>=2.0.0
